+++ src/main/java/io/github/a5b84/helditeminfo/mixin/InGameHudAccessor.java
@@ -7,6 +7,11 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 @Mixin(InGameHud.class)
 public interface InGameHudAccessor {
 
+    @Accessor("field_32169")
+    static int getNumHeartsPerRow() {
+        throw new AssertionError();
+    }
+
     @Accessor("field_32170")
     static int getLineHeight() {
         throw new AssertionError();