+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 - Updated to 1.21.6
-- Overlay messages are now moved up when they would overlap with the tooltip ([#36](https://github.com/A5b84/held-item-info/issues/36))
+- The tooltip and overlay messages are now moved up when they would overlap with each other or with the health bar ([#36](https://github.com/A5b84/held-item-info/issues/36))
 
 ### [1.7.2] (1.21.5)
 - Updated to 1.21.5 ([#37](https://github.com/A5b84/held-item-info/pull/37))